rework wiki page widget
