+++ src/main/java/org/sagebionetworks/web/client/widget/entity/EntityPageTopViewImpl.java
@@ -352,7 +352,7 @@ public class EntityPageTopViewImpl extends Composite implements EntityPageTopVie
 	private void addWikiPageWidget(LayoutContainer container, EntityBundle bundle, boolean canEdit, int spanWidth) {
 		if (DisplayUtils.isWikiSupportedType(bundle.getEntity())) {
 			// Child Page Browser
-			container.add(wikiPageWidget);
+			container.add(wikiPageWidget.asWidget());
 			wikiPageWidget.configure(new WikiPageKey(bundle.getEntity().getId(), WidgetConstants.WIKI_OWNER_ID_ENTITY, null), canEdit, new WikiPageWidget.Callback() {
 				@Override
 				public void pageUpdated() {
