+++ src/main/java/org/sagebionetworks/web/client/widget/entity/MarkdownWidget.java
@@ -9,7 +9,6 @@ import org.sagebionetworks.web.client.IconsImageBundle;
 import org.sagebionetworks.web.client.SynapseClientAsync;
 import org.sagebionetworks.web.client.SynapseJSNIUtils;
 import org.sagebionetworks.web.client.widget.WidgetRendererPresenter;
-import org.sagebionetworks.web.client.widget.entity.registration.WidgetConstants;
 import org.sagebionetworks.web.client.widget.entity.registration.WidgetRegistrar;
 import org.sagebionetworks.web.shared.WikiPageKey;
 
@@ -63,7 +62,7 @@ public class MarkdownWidget extends LayoutContainer {
 					add(panel);
 					layout();
 					synapseJSNIUtils.highlightCodeBlocks();
-					MarkdownWidget.loadTableSorters(panel, synapseJSNIUtils);
+					DisplayUtils.loadTableSorters(panel, synapseJSNIUtils);
 					//asynchronously load the widgets
 					loadWidgets(panel, wikiKey, isWiki, widgetRegistrar, synapseClient, iconsImageBundle, isPreview);
 				} catch (JSONObjectAdapterException e) {
@@ -116,17 +115,6 @@ public class MarkdownWidget extends LayoutContainer {
 			i++;
 			currentWidgetDiv = DisplayConstants.DIV_ID_WIDGET_PREFIX + i + suffix;
 			el = panel.getElementById(currentWidgetDiv);
-		}
-	}
-	
-	public static void loadTableSorters(final HTMLPanel panel, SynapseJSNIUtils synapseJSNIUtils) {
-		String id = WidgetConstants.MARKDOWN_TABLE_ID_PREFIX;
-		int i = 0;
-		Element table = panel.getElementById(id + i);
-		while (table != null) {
-			synapseJSNIUtils.tablesorter(id+i);
-			i++;
-			table = panel.getElementById(id + i);
 		}
 	}
